requirements.txt
# Core LLM and Graph
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-aws>=0.1.0
boto3>=1.34.0

# CLI and utilities
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
